+++ include/client/preload_util.hpp
@@ -32,7 +32,7 @@
 
 #include <client/preload.hpp>
 #include <common/metadata.hpp>
-
+#include <optional>
 #include <string>
 #include <iostream>
 #include <map>