+++ include/daemon/classes/fs_data.hpp
@@ -34,6 +34,7 @@
 #include <unordered_map>
 #include <map>
 #include <functional> //std::hash
+#include <string_view>
 
 /* Forward declarations */
 namespace gkfs {
@@ -69,6 +70,12 @@ private:
 
     // Database
     std::shared_ptr<gkfs::metadata::MetadataDB> mdb_;
+    std::string dbbackend_;
+    // Kreon
+    bool reuse_md_ = false;
+    bool keep_md_ = false;
+    unsigned long long size_md_ = 8589934592ull;
+
     // Storage backend
     std::shared_ptr<gkfs::data::ChunkStorage> storage_;
 
@@ -122,6 +129,12 @@ public:
 
     void
     metadir(const std::string& metadir_);
+
+    std::string_view
+    dbbackend() const;
+
+    void
+    dbbackend(const std::string& dbbackend_);
 
     const std::shared_ptr<gkfs::metadata::MetadataDB>&
     mdb() const;
@@ -191,6 +204,24 @@ public:
 
     void
     blocks_state(bool blocks_state);
+
+    bool
+    kreon_keep_md() const;
+
+    void
+    kreon_keep_md(bool keep_md);
+
+    bool
+    kreon_reuse_md() const;
+
+    void
+    kreon_reuse_md(bool reuse_md);
+
+    unsigned long long
+    kreon_size_md() const;
+
+    void
+    kreon_size_md(unsigned int size_md);
 };
 
 } // namespace daemon