+++ src/daemon/classes/fs_data.cpp
@@ -110,6 +110,16 @@ FsData::metadir(const std::string& metadir) {
     FsData::metadir_ = metadir;
 }
 
+std::string_view
+FsData::dbbackend() const {
+    return dbbackend_;
+}
+
+void
+FsData::dbbackend(const std::string& dbbackend) {
+    FsData::dbbackend_ = dbbackend;
+}
+
 const std::string&
 FsData::rpc_protocol() const {
     return rpc_protocol_;
@@ -198,6 +208,37 @@ FsData::blocks_state() const {
 void
 FsData::blocks_state(bool blocks_state) {
     FsData::blocks_state_ = blocks_state;
+}
+
+bool
+FsData::kreon_keep_md() const {
+    return keep_md_;
+}
+
+void
+FsData::kreon_keep_md(bool keep_md) {
+    FsData::keep_md_ = keep_md;
+}
+
+bool
+FsData::kreon_reuse_md() const {
+    return reuse_md_;
+}
+
+void
+FsData::kreon_reuse_md(bool reuse_md) {
+    FsData::reuse_md_ = reuse_md;
+}
+
+unsigned long long
+FsData::kreon_size_md() const {
+    return size_md_;
+}
+
+void
+FsData::kreon_size_md(unsigned int size_md) {
+    FsData::size_md_ =
+            (unsigned long long) size_md * 1024ull * 1024ull * 1024ull;
 }
 
 } // namespace gkfs::daemon